New dir for computer vision. moved related part
