+++ client/computer_vision/color_object_recognition.h
