+++ client/computer_vision/imshow_queue.h
